requirements.txt
customtkinter==5.2.2
minecraft-launcher-lib==8.0
pillow==10.4.0
